+++ Assets/EditorTests/Unit/Components/Generators.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c390271d3f623415f8c0503659de5bad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
