+++ Assets/EditorTests/Unit/Components/Generators/PlaneTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf1471db0eb30454d83b22dd3dac2655
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
