+++ Packages/com.daniil-ryzhkov.proceduraltoolkit/Components/Generators/BaseGeneratorComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e3135401aacf14a1da155f009b2faedc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
