+++ Packages/com.daniil-ryzhkov.proceduraltoolkit/Components/Generators/IGeneratorComponent.cs
@@ -1,0 +1,10 @@
+using System;
+using ProceduralToolkit.Services.Generators;
+
+namespace ProceduralToolkit.Components.Generators
+{
+    public interface IGeneratorComponent : IGenerator
+    {
+        event Action GeneratorUpdated;
+    }
+}
