+++ Packages/com.daniil-ryzhkov.proceduraltoolkit/Components/Generators/IGeneratorComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c52456bd8d8a84c4d93814d84ee8c7c0
+guid: 1c67074744d6e4f279317959fa0af98b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
