+++ Packages/com.daniil-ryzhkov.proceduraltoolkit/Components/Generators/Plane.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using ProceduralToolkit.Models;
 using ProceduralToolkit.Services.DI;
 using ProceduralToolkit.Services.Generators;
@@ -7,7 +6,7 @@ using UnityEngine;
 
 namespace ProceduralToolkit.Components.Generators
 {
-    public class Plane : MonoBehaviour, IGenerator
+    public class Plane : BaseGeneratorComponent
     {
         [Min(0)]
         public float length;
@@ -17,42 +16,19 @@ namespace ProceduralToolkit.Components.Generators
         [Service]
         private readonly Func<PlaneGeneratorSettings, IGenerator> planeGeneratorProvider;
 
-        private IGenerator Generator
+        protected override IGenerator UpdateGenerator()
         {
-            get
-            {
-                if (generator == null)
-                {
-                    UpdateGenerator();
-                }
-                return generator;
-            }
-            set => generator = value;
-        }
-        private IGenerator generator;
-
-        private void UpdateGenerator()
-        {
-            Generator = planeGeneratorProvider?.Invoke(new PlaneGeneratorSettings()
+            return planeGeneratorProvider?.Invoke(new PlaneGeneratorSettings()
             {
                 Length = length,
                 Width = width
             });
         }
 
-        public IEnumerable<Vector3> Vertices => Generator.Vertices;
-
-        public IEnumerable<int> Triangles => Generator.Triangles;
-
-        public void Reset()
+        private void Reset()
         {
             length = 1;
             width = 1;
-        }
-
-        public void OnValidate()
-        {
-            UpdateGenerator();
         }
     }
 }
